refactors systems into system_XXX files
