+++ mescal/visualizations/folium_viz_system/folium_generators.py
@@ -1,20 +1,10 @@
 from abc import ABC, abstractmethod
-from typing import Callable, TYPE_CHECKING, Type
-import hashlib
+from typing import TYPE_CHECKING
 
 import folium
-from folium.plugins import PolyLineTextPath, PolyLineOffset, AntPath
-from shapely import Polygon, MultiPolygon, LineString
 
-from mescal.visualizations.folium_viz_system.folium_styling import (
-    AreaStyleResolver,
-    LineStyleResolver,
-    CircleMarkerStyleResolver,
-    TextOverlayStyleResolver,
-    LineTextStyleResolver,
-)
 from mescal.visualizations.folium_viz_system.system_base import ResolvedStyle, StyleResolver, FoliumObjectGenerator
-from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem, KPIDataItem
+from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem
 from mescal.utils.logging import get_logger
 
 if TYPE_CHECKING:
@@ -65,231 +55,6 @@ class IconGenerator(ABC):
         pass
 
 
-class AreaGenerator(FoliumObjectGenerator[AreaStyleResolver]):
-    """Generates folium GeoJson objects for area geometries."""
-
-    def _style_resolver_type(self) -> Type[AreaStyleResolver]:
-        return AreaStyleResolver
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        geometry = data_item.get_geometry()
-        if not isinstance(geometry, (Polygon, MultiPolygon)):
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-        popup = self.popup_generator.generate_popup(data_item) if self.popup_generator else None
-
-        style_dict = {
-            'fillColor': style.fill_color,
-            'color': style.border_color,
-            'weight': style.border_width,
-            'fillOpacity': style.fill_opacity
-        }
-
-        highlight_dict = style_dict.copy()
-        highlight_dict['weight'] = style.highlight_border_width
-        highlight_dict['fillOpacity'] = style.highlight_fill_opacity
-
-        geojson_data = {
-            "type": "Feature",
-            "geometry": geometry.__geo_interface__,
-            "properties": {"tooltip": tooltip}
-        }
-
-        geojson_kwargs = {
-            'style_function': lambda x, s=style_dict: s,
-            'highlight_function': lambda x, h=highlight_dict: h,
-            'tooltip': folium.GeoJsonTooltip(fields=['tooltip'], aliases=[''], sticky=True)
-        }
-
-        if popup:
-            geojson_kwargs['popup'] = folium.Popup(popup, max_width=300)
-
-        folium.GeoJson(geojson_data, **geojson_kwargs).add_to(feature_group)
-
-
-class LineGenerator(FoliumObjectGenerator[LineStyleResolver]):
-    """Generates folium PolyLine objects for line geometries with optional per-feature-group offset tracking."""
-
-    def __init__(
-            self,
-            style_resolver: LineStyleResolver = None,
-            tooltip_generator: TooltipGenerator = None,
-            popup_generator: PopupGenerator = None,
-            per_feature_group_offset_registry: bool = True,
-            offset_increment: int = 5,
-    ):
-        super().__init__(style_resolver, tooltip_generator, popup_generator)
-        self.offset_increment = offset_increment
-        self.per_feature_group_registry = per_feature_group_offset_registry
-        self._global_registry: dict[str, int] = {}
-        self._group_registry: dict[int, dict[str, int]] = {}
-
-    def reset_registry(self) -> None:
-        self._global_registry.clear()
-        self._group_registry.clear()
-
-    def _style_resolver_type(self) -> Type[LineStyleResolver]:
-        return LineStyleResolver
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        geometry = data_item.get_geometry()
-        if not isinstance(geometry, LineString):
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-        popup = self.popup_generator.generate_popup(data_item) if self.popup_generator else None
-
-        coordinates = [(lat, lon) for lon, lat in geometry.coords]
-        effective_offset = self._get_line_offset(coordinates, feature_group)
-
-        if style.reverse_path_direction:
-            coordinates = coordinates[::-1]
-
-        line_kwargs = dict(
-            locations=coordinates,
-            color=style.line_color,
-            weight=style.line_width,
-            opacity=style.line_opacity,
-            offset=effective_offset,
-            tooltip=tooltip,
-            dash_array=style.dash_pattern or None,
-            popup=folium.Popup(popup, max_width=300) if popup else None,
-        )
-
-        if style.line_ant_path:
-            if effective_offset not in [None, 0, 0.0]:
-                logger.warning(
-                    f'Trying to set an offset with an animated LineAntPath is not possible. '
-                    f'Offset will not be applied for {data_item.get_name()}.'
-                )
-            line_kwargs.update(dict(
-                paused=False,
-                reverse=False,
-                hardware_acceleration=False,
-                delay=style.line_ant_path_delay,
-                pulse_color=style.line_ant_path_pulse_color,
-            ))
-            poly_line = AntPath(**line_kwargs)
-        else:
-            poly_line = PolyLineOffset(**line_kwargs)
-
-        poly_line.add_to(feature_group)
-
-    def _get_line_offset(self, coordinates: list[tuple[float, float]], feature_group: folium.FeatureGroup) -> float:
-        line_hash = self._hash_coordinates(coordinates)
-        registry = self._get_registry_for_group(feature_group)
-        offset_index = registry.get(line_hash, 0)
-        registry[line_hash] = offset_index + 1
-        offset_px = offset_index * self.offset_increment
-        offset_side = 1 if offset_index % 2 == 0 else -1
-        effective_offset = offset_px * offset_side
-        return effective_offset
-
-    def _get_registry_for_group(self, feature_group: folium.FeatureGroup) -> dict[str, int]:
-        if not self.per_feature_group_registry:
-            return self._global_registry
-        group_id = id(feature_group)
-        if group_id not in self._group_registry:
-            self._group_registry[group_id] = {}
-        return self._group_registry[group_id]
-
-    def _hash_coordinates(self, coordinates: list[tuple[float, float]]) -> str:
-        rounded = [(round(lat, 6), round(lon, 6)) for lat, lon in coordinates]
-        return hashlib.md5(str(rounded).encode("utf-8")).hexdigest()
-
-
-class LineTextOverlayGenerator(FoliumObjectGenerator[LineTextStyleResolver]):
-    def __init__(
-            self,
-            style_resolver: LineTextStyleResolver = None,
-            tooltip_generator=None,
-            popup_generator=None,
-            text_formatter: Callable[[MapDataItem], str] = None
-    ):
-        super().__init__(style_resolver or LineTextStyleResolver(), tooltip_generator, popup_generator)
-        self.text_formatter = text_formatter or self._default_text_formatter
-
-    def _default_text_formatter(self, data_item: MapDataItem) -> str:
-        if isinstance(data_item, KPIDataItem):
-            # TODO: use quantity and pretty value; auto-num-of-decimals and so on...
-            return f"{data_item.kpi.value:.1f}"
-        return str(data_item.object_id)
-
-    def _style_resolver_type(self) -> Type[LineTextStyleResolver]:
-        return LineTextStyleResolver
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        geometry = data_item.get_geometry()
-        if not isinstance(geometry, LineString):
-            return
-
-        text = self.text_formatter(data_item)
-        if not text:
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        coordinates = [(lat, lon) for lon, lat in geometry.coords]
-
-        if style.reverse_path_direction:
-            coordinates = coordinates[::-1]
-
-        invisible_line = folium.PolyLine(
-            locations=coordinates,
-            color=None,
-            opacity=0.0,
-        )
-        invisible_line.add_to(feature_group)
-
-        PolyLineTextPath(
-            invisible_line,
-            text=text,
-            repeat=style.text_repeat,
-            center=style.text_center,
-            below=style.text_below,
-            orientation=style.text_orientation,
-            offset=style.text_offset,
-            attributes={
-                'font-weight': style.font_weight,
-                'font-size': str(style.font_size),
-                'fill': style.font_color,
-            }
-        ).add_to(feature_group)
-
-
-class CircleMarkerGenerator(FoliumObjectGenerator[CircleMarkerStyleResolver]):
-    """Generates folium CircleMarker objects for point geometries."""
-
-    def _style_resolver_type(self) -> Type[CircleMarkerStyleResolver]:
-        return CircleMarkerStyleResolver
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        try:
-            location = data_item.get_location()
-        except ValueError:
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-        popup = self.popup_generator.generate_popup(data_item) if self.popup_generator else None
-
-        marker_kwargs = {'location': location, 'tooltip': tooltip}
-        if popup:
-            marker_kwargs['popup'] = folium.Popup(popup, max_width=300)
-
-        circle_kwargs = {
-            'radius': style.radius,
-            'color': style.border_color,
-            'fillColor': style.fill_color,
-            'fillOpacity': style.fill_opacity,
-            'weight': style.border_width,
-            **marker_kwargs
-        }
-        folium.CircleMarker(**circle_kwargs).add_to(feature_group)
-
-
 class _TMPIconGenerator(FoliumObjectGenerator, ABC):
     """Generates folium Marker or CircleMarker objects for point geometries."""
 
@@ -326,94 +91,3 @@ class _TMPIconGenerator(FoliumObjectGenerator, ABC):
         folium.Marker(**marker_kwargs).add_to(feature_group)
 
 
-class TextOverlayGenerator(FoliumObjectGenerator[TextOverlayStyleResolver]):
-    """Generates text overlays for map data items."""
-
-    def __init__(
-            self,
-            style_resolver: TextOverlayStyleResolver = None,
-            tooltip_generator: TooltipGenerator = None,
-            popup_generator: PopupGenerator = None,
-            text_formatter: Callable[[MapDataItem], str] = None
-    ):
-        super().__init__(style_resolver or TextOverlayStyleResolver(), tooltip_generator, popup_generator)
-        self.text_formatter = text_formatter or self._default_text_formatter
-
-    def _style_resolver_type(self) -> Type[TextOverlayStyleResolver]:
-        return TextOverlayStyleResolver
-
-    def _default_text_formatter(self, data_item: MapDataItem) -> str:
-        if isinstance(data_item, KPIDataItem):
-            # TODO: use quantity and pretty value; auto-num-of-decimals and so on...
-            return f"{data_item.kpi.value:.1f}"
-        return str(data_item.object_id)
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        try:
-            location = data_item.get_location()
-        except ValueError:
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        text = self.text_formatter(data_item)
-        popup = self.popup_generator.generate_popup(data_item) if self.popup_generator else None
-
-        text_color = style.text_color
-        font_size = style.font_size
-        font_weight = style.font_weight
-        shadow_size = style.shadow_size
-        shadow_color = style.shadow_color
-
-        icon_html = f'''
-            <div style="
-                position: absolute;
-                left: 50%;
-                top: 50%;
-                transform: translate(-50%, -50%);
-                text-align: center;
-                font-size: {font_size};
-                font-weight: {font_weight};
-                color: {text_color};
-                white-space: nowrap;
-                text-shadow:
-                   -{shadow_size} -{shadow_size} 0 {shadow_color},  
-                    {shadow_size} -{shadow_size} 0 {shadow_color},
-                   -{shadow_size}  {shadow_size} 0 {shadow_color},
-                    {shadow_size}  {shadow_size} 0 {shadow_color};
-            ">
-                {text}
-            </div>
-        '''
-
-        marker_kwargs = {
-            'location': location,
-            'icon': folium.DivIcon(html=icon_html)
-        }
-
-        if popup:
-            marker_kwargs['popup'] = folium.Popup(popup, max_width=300)
-
-        folium.Marker(**marker_kwargs).add_to(feature_group)
-
-    def _get_contrasting_color(self, surface_color: str) -> str:
-        """Get contrasting text color for a surface color."""
-        if self._is_dark(surface_color):
-            return '#F2F2F2'
-        return '#3A3A3A'
-
-    def _get_shadow_color(self, text_color: str) -> str:
-        """Get shadow color for text."""
-        if text_color == '#F2F2F2':
-            return '#3A3A3A'
-        return '#F2F2F2'
-
-    @staticmethod
-    def _is_dark(color: str) -> bool:
-        """Check if a color is dark."""
-        if not color.startswith('#'):
-            return False
-        try:
-            r, g, b = [int(color[i:i + 2], 16) for i in (1, 3, 5)]
-            return (0.299 * r + 0.587 * g + 0.114 * b) < 160
-        except (ValueError, IndexError):
-            return False
